+++ tests/testthat/test-dev_ermod_lin.R
@@ -209,10 +209,17 @@ mod_mtcars <- dev_ermod_bin(
 
 test_that("Exposure metrics selection", {
   expect_equal(extract_var_exposure(ermod_bin_exp_sel), "AUCss_1000")
-  expect_equal(
-    extract_exp_sel_comp(ermod_bin_exp_sel)[, 1],
-    c(AUCss_1000 = 0.000000, Cmaxss = -0.6490312, Cminss = -2.7563957)
-  )
+  if ("model" %in% colnames(extract_exp_sel_comp(ermod_bin_exp_sel))) {
+    expect_equal(
+      extract_exp_sel_comp(ermod_bin_exp_sel)[, 2],
+      c(0.000000, -0.6490312, -2.7563957)
+    )
+  } else {
+    expect_equal(
+      extract_exp_sel_comp(ermod_bin_exp_sel)[, 1],
+      c(AUCss_1000 = 0.000000, Cmaxss = -0.6490312, Cminss = -2.7563957)
+    )
+  }
 
   priors <- prior_summary(ermod_bin_exp_sel$l_mod_exposures[[2]])
   expect_equal(priors$prior$scale, 5)